backend/requirements.txt
Django==3.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
django-cors-headers==3.13.0
django-nose==1.4.6
coverage==5.0.4
